啊啊啊啊啊啊啊啊啊啊啊啊战斗系统完成啦啊啊啊啊啊啊啊

+++ Assets/TeamController.cs
@@ -29,8 +29,8 @@ public class TeamController : MonoBehaviour
     }
     [System.Serializable]
     public struct Member{
-        public int HP, MaxHP, MP, MaxMP;
-        public int ATK, DEF, SPD;
+        public float HP, MaxHP, MP, MaxMP;
+        public float ATK, DEF, SPD;
         public JOB[] Job;
         public string Name;
         public List<string> Equipments;
